requirements.txt
amqp==2.5.2
astroid==2.3.3
autopep8==1.5
bcrypt==3.1.7
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
Django==2.2.10
django-celery-beat==1.6.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-mysql==3.3.0
django-redis==4.11.0
django-timezone-field==3.1
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
drf-yasg==1.17.0
entrypoints==0.3
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==4.0.0
flake8==3.7.9
geographiclib==1.50
geopy==1.21.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
kombu==4.6.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.18.1
openpyxl==3.0.3
packaging==20.1
pandas==1.0.1
phonenumbers==8.11.4
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
text-unidecode==1.3
twilio==6.35.4
typed-ast==1.4.1
unicode==2.7
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==2.2.0
django-rest-swagger==2.1.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
django-storages==1.10.1
boto3==1.15.0